requirements.txt
# Core AI/ML Libraries
langchain==0.1.0
langchain-openai==0.0.5
langchain-community==0.0.10
openai==1.6.1

# Vector Database
chromadb==0.4.22
faiss-cpu==1.7.4

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
streamlit==1.28.1

# HTTP Client
httpx==0.25.2
requests==2.31.0

# Data Processing
pandas==2.1.4
numpy==1.24.3
pydantic==2.5.2

# Document Processing
pypdf==3.17.4
python-docx==1.1.0
beautifulsoup4==4.12.2
markdown==3.5.1

# Configuration & Environment
python-dotenv==1.0.0
pyyaml==6.0.1

# Logging & Monitoring
structlog==23.2.0
prometheus-client==0.19.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0

# Code Quality
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Security
cryptography==41.0.8
passlib==1.7.4

# Email
aiosmtplib==3.0.1
email-validator==2.1.0

# Async Support
asyncio-mqtt==0.16.1
websockets==12.0

# Utilities
click==8.1.7
rich==13.7.0
tqdm==4.66.1 